+++ ReportTree.Server/DTOs/AuditDtos.cs
@@ -1,0 +1,8 @@
+using ReportTree.Server.Models;
+
+namespace ReportTree.Server.DTOs;
+
+public class AuditExportQueryParameters : AuditLogQuery
+{
+    public string Format { get; set; } = "csv";
+}